Merge the NumSciPy package with my local repository.  This represents several (minor) bug fixes and several (major) functional additions:
ArrayIO.py:
Adds some IO functionality to the array library.  Specifically, the 
ability to read RGBA PNG images.  

ReadPNG - Read a PNG image into a Numpy Array data object.  This module 
is loaded iff the pylab package is present.

Imaging.py:
Adds some image processing functionality native to the numpy and scipy 
distribution through scipy.ndimage.  Wrapped functionality includes:

ExtractRGBAChannel - Extracts a single channel from an array 
representing an RGBA image

GaussianGradientMagnitude - Computes a multi-dimensional gradient 
magnitude dataset from a multi-dimensional input array.  If the standard 
deviation of the input gaussians is a single number, it behaves as an 
isotropic kernel.  If a list of sigmas is given, the kernel is built as 
separable gaussians with the appropriate sigmas in the order they are 
given.

JointHistogram - Computes a joint histogram from two mutli-dimensional 
arrays.  The input arrays must be of the same size and dimension, but 
these specifications are arbitrary.  The port Array One specifies the 
array to be considered as the 'x axis' in the histogram with the other 
being the y axis.  

GaussianSmooth - Smooths the input array with a multi-dimensional 
gaussian kernel.  The kernel is specified in the same way as the 
GaussianGradientMagnitude where the standard deviation, sigma, is given 
as either a single element (isotropic kernel) or as a list.

MedianFilter - Smooths the input array with a multi-dimensional median 
filter.  The filter is specified the same way as the GaussianSmooth 
filter only instead of sigma, the size (footprint) of the filter is 
given as either a single element or a list.

ImageDifference - Compute the difference between two input arrays 
representing equally sized images.

ImageNormalize - Normalize the input image to bring the range of values 
it contains to [0,1].

ArrayAccess.py:
Contains bug-fixes found in indexing and transposition of arrays

ArrayConvert.py:
Added support for arrays to be converted to vtk-usable datatypes and 
SciPy Matrices.  Converted base class to subclass from object to prevent 
mix-in bugs.

ArrayToMatrix -  Converts a Numpy Array datatype to a SciPy Matrix type.  
This works for arrays and matrices up to and including rank 3.

ArrayToVTKImageData - Convert a Numpy Array datatype to a vtkImageData 
type.  This works for arrays up to and including rank 3.

ArrayOperations.py:
Adds functionality to manipulate arrays.

ArrayScalarMultiply - Atomic operation multiplying all values in a Numpy 
Array with a single scalar value.

ArrayExtractRegion - Extract a sub-array from the given array.  This can 
be used to extract a lower-dimensional portion of the array or a small, 
contiguous sub-volume of a dataset.

DSP.py:
Added support for ensemble processing in all existing methods.

MatrixUtilities.py:
Refactored Matlab Reader to increase flexibility.

MatrixWriter - Added support for writing a numpy array or scipy matrix 
to a Matlab .mat file.

__init__.py:
Added required class registrations.