Commiting many changes that I did in the trunk...
CONTROL FLOW
- packages/controlflow/fold.py
Instead of using ModuleError, now the VistrailsInternalError is being 
used, once the error handling is inside updateUpstream() method;
Type checking was added inside Fold (it verifies if each element of the 
list has the same type as the input port chosen by the user); this type 
checking is done before the computation starts; only boolean, string, 
integer, float, file, list and tuple values are verified; if the type 
can not be identified, the type checking is simply ignored.
New methods:
updateFunctionPort(): it was inside updateUpstream; now it's a method of 
the Fold module;
setInputValues(): used to set a value in an input port;
typeChecking(): it does the type checking;
createSignature(): used to create a signature for a port spec (used 
inside typeChecking());
compare(): used to compare two port specs, to see if they match (used 
inside typeChecking()).

- packages/controlflow/conditional.py
Error handling was added in some parts, using VistrailsInternalError.

- packages/controlflow/products.py
Just a better organization of the code...

- packages/controlflow/utils.py
Just a better organization of the code...

EXECUTION LOG
- db/versions/v0_9_5/schemas/sql/vistrails_drop.sql
- db/versions/v0_9_5/schemas/sql/vistrails.sql
- db/versions/v0_9_5/persistence/sql/auto_gen.py
- db/versions/v0_9_5/persistence/xml/auto_gen.py
- db/versions/v0_9_5/domain/auto_gen.py
- db/versions/v0_9_5/specs/all.xml
- db/specs/all.xml
group_exec and loop_exec were added, in order to keep track of
group/subworkflow and loop information.

- core/log/group_exec.py
A new file added.

- core/log/workflow_exec.py
New method to deal with its child nodes (module_exec and group_exec).

- core/log/loop_exec.py
New methods to deal with its child nodes (module_exec and group_exec).

- core/log/controller.py
New methods were created so the group/subworkflow and loop information 
could be written in the execution log.
New methods:
create_module_exec(): creates an object of ModuleExec;
create_group_exec(): creates an object of GroupExec;
create_loop_exec(): creates an object of LoopExec;
start_module_loop_execution(): adds information of a module execution to 
the log, if it has a loop operation;
finish_module_loop_execution();
start_group_loop_execution(): adds information of a group/subworkflow 
execution to the log, if it has a loop operation;
finish_group_loop_execution();
start_group_execution(): adds information of a group/subworkflow 
execution to the log;
finish_group_execution();
start_loop_execution(): adds information of a loop execution to the log;
finish_loop_execution();
start_execution(): the main method to add information to the log;
finish_execution().

- core/interpreter/cached.py
Instead of calling start_module_execution and finish_module_execution 
inside execute_pipeline(), now the methods start_execution and 
finish_execution are called; the log of the workflow is now stored 
inside the logging_obj attribute of a module, so the modules inside a 
group/subworkflow can use it to write its own information; in 
finalize_pipeline() method, the flag 'computed' is set to False for all 
modules of the pipeline.

- core/modules/vistrails_module.py
A flag 'computed' was added to the Module class; with this flag, a 
module that was already computed will not be stored again inside the 
execution log.

- core/modules/sub_module.py
Now, the log of the workflow and the group_exec of a group is passed to 
its internal modules.

OTHER MODIFICATIONS
- core/modules/sub_module.py
ModuleError now handles the errors that occur inside a group.

- core/modules/basic_modules.py
The output port 'self' is now optional for all modules, except 
PythonSource.