Added example of enumeration module (and widget) to the SAHM package.
We may eventually move this to the core VisTrails modules, but this is
currently set up as an example for Colin.  It allows a user to define
an enumeration module by subclassing for the constant class that
serves as the type of the enumeration, and building an enumeration
widget by passing the possible values.

packages/sahm/enum_modules.py:
  - MyEnum: new example module

packages/sahm/enum_widgets.py:
  - EnumWidget: new class, base class for all enumeration widgets
  - build_enum_widget: dynamically creates a subclass of EnumWidget
    and sets the widget's possible values